Update app.spec.js
Corrected nested it( function (t('it should pass if "getTimeZone" filter returns timezone from given timestamp', function ()) that appeared after the line 61 that appeared after a line with the same text. This marked a build error saying it( should be only in a describe, studying the function it of mocha and other function calls, the nested it function should be a mistake.